fix: refactor to complete all the routes for each dependency
